Move view to separate file [skip ci]
